Removed All Previous Models References
